+++ src/app/feature/product/models/product.model.ts
@@ -1,13 +1,14 @@
 import { Size } from './Size.enum';
 import { IProduct } from './product.interface';
 
-export class ProductModel implements IProduct {
+export class ProductModel implements IProduct { 
   constructor(
     public name: string,
     public description: string,
     public price: number,
     public category: Size = Size.unknown,
-    public isAvailable: boolean = false) {
-
-  }
+    public count:number)
+     {  }
+  get isAvailable(){
+    return this.count > 0 };
 }