+++ src/app/feature/shared/pipes/total.pipe.ts
@@ -1,0 +1,7 @@
+/*import { Pipe, PipeTransform } from '@angular/core';
+import { ProductModel } from '../../product/models/product.model';
+
+@Pipe({name: 'total'})
+export class TotalSumPipe implements PipeTransform{
+    transform(value:ProductModel)
+}*/